various: move sorting away from config
